Adding doc for transient dismissal logic
